feat(forest fire): save mouse pos in integer vector2

+++ hello-open-gl/src/scenes/SceneForestFire.h
@@ -37,6 +37,6 @@ namespace scene {
         float m_GrowthProbability;
 
         bool m_IsMousePressed;
-        glm::vec2 m_MousePosition;
+        glm::ivec2 m_MousePosition;
     };
 }
